backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai

# langchain==0.3.13
# langchain-core==0.3.28
# langchain-community==0.3.13
# langchain-openai==0.2.14
